+++ benchmark/xor/xor_test.go
@@ -355,7 +355,7 @@ func TestRunPredictiveWithPlasticity(t *testing.T) {
 }
 
 func TestWeightStats(t *testing.T) {
-	weights := []int16{100, 200, 300, 400, 500}
+	weights := []int32{100, 200, 300, 400, 500}
 	mean, stddev := benchmark.WeightStats(weights)
 
 	if mean != 300 {
